Rename package to testgen
