+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/testgen/HeadCuttingMutator.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.optaplanner.core.impl.score.director.drools.reproducer;
+package org.optaplanner.core.impl.score.director.drools.testgen;
 
 import java.util.List;
 
-import org.optaplanner.core.impl.score.director.drools.reproducer.operation.KieSessionOperation;
+import org.optaplanner.core.impl.score.director.drools.testgen.operation.KieSessionOperation;
 
 class HeadCuttingMutator {
