+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/testgen/NoopJournal.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.optaplanner.core.impl.score.director.drools.reproducer;
+package org.optaplanner.core.impl.score.director.drools.testgen;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -21,9 +21,9 @@ import java.util.List;
 
 import org.kie.api.runtime.KieSession;
 import org.optaplanner.core.impl.domain.variable.descriptor.VariableDescriptor;
-import org.optaplanner.core.impl.score.director.drools.reproducer.fact.Fact;
-import org.optaplanner.core.impl.score.director.drools.reproducer.operation.KieSessionInsert;
-import org.optaplanner.core.impl.score.director.drools.reproducer.operation.KieSessionOperation;
+import org.optaplanner.core.impl.score.director.drools.testgen.fact.Fact;
+import org.optaplanner.core.impl.score.director.drools.testgen.operation.KieSessionInsert;
+import org.optaplanner.core.impl.score.director.drools.testgen.operation.KieSessionOperation;
 
 class NoopJournal implements KieSessionJournal {
