+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/testgen/fact/ExistingInstanceValueProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.optaplanner.core.impl.score.director.drools.reproducer.fact;
+package org.optaplanner.core.impl.score.director.drools.testgen.fact;
 
 class ExistingInstanceValueProvider extends AbstractValueProvider {
