+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/drools/testgen/operation/KieSessionDelete.java
@@ -13,35 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.optaplanner.core.impl.score.director.drools.reproducer.operation;
+package org.optaplanner.core.impl.score.director.drools.testgen.operation;
 
 import org.kie.api.runtime.KieSession;
-import org.optaplanner.core.impl.score.director.drools.reproducer.fact.Fact;
+import org.optaplanner.core.impl.score.director.drools.testgen.fact.Fact;
 import org.slf4j.Logger;
 
-public class KieSessionInsert implements KieSessionOperation {
+public class KieSessionDelete implements KieSessionOperation {
 
     private final int id;
-    private final Fact fact;
+    private final Fact entity;
 
-    public KieSessionInsert(int id, Fact fact) {
+    public KieSessionDelete(int id, Fact entity) {
         this.id = id;
-        this.fact = fact;
-    }
-
-    public Fact getFact() {
-        return fact;
+        this.entity = entity;
     }
 
     @Override
     public void invoke(KieSession kieSession) {
-        kieSession.insert(fact.getInstance());
+        kieSession.delete(kieSession.getFactHandle(entity.getInstance()));
     }
 
     @Override
     public void print(Logger log) {
         log.debug("        //{}", this);
-        log.info("        kieSession.insert({});", fact);
+        log.info("        kieSession.delete(kieSession.getFactHandle({}), {});", entity, entity);
     }
 
     @Override
